+++ src/main/java/org/jabref/gui/groups/GroupDialogViewModel.java
@@ -33,6 +33,7 @@ import org.jabref.logic.util.StandardFileType;
 import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.database.BibDatabaseContext;
+import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.Keyword;
 import org.jabref.model.entry.field.FieldFactory;
 import org.jabref.model.groups.AbstractGroup;
@@ -76,6 +77,7 @@ public class GroupDialogViewModel {
     private final BooleanProperty typeSearchProperty = new SimpleBooleanProperty();
     private final BooleanProperty typeAutoProperty = new SimpleBooleanProperty();
     private final BooleanProperty typeTexProperty = new SimpleBooleanProperty();
+    private final BooleanProperty typeSelectionProperty = new SimpleBooleanProperty();
 
     // Option Groups
     private final StringProperty keywordGroupSearchTermProperty = new SimpleStringProperty("");
@@ -94,6 +96,7 @@ public class GroupDialogViewModel {
     private final StringProperty autoGroupPersonsFieldProperty = new SimpleStringProperty("");
 
     private final StringProperty texGroupFilePathProperty = new SimpleStringProperty("");
+    private final BooleanProperty entriesAreSelected = new SimpleBooleanProperty(false);
 
     private Validator nameValidator;
     private Validator nameContainsDelimiterValidator;
@@ -388,6 +391,14 @@ public class GroupDialogViewModel {
                         new DefaultAuxParser(new BibDatabase()),
                         fileUpdateMonitor,
                         currentDatabase.getMetaData());
+            } else if (typeSelectionProperty.getValue()) {
+                ExplicitGroup tempResultingGroup = new ExplicitGroup(
+                        groupName,
+                        groupHierarchySelectedProperty.getValue(),
+                        preferencesService.getBibEntryPreferences().getKeywordSeparator()
+                );
+                tempResultingGroup.add(selectedEntries);
+                resultingGroup = tempResultingGroup;
             }
 
             if (resultingGroup != null) {
@@ -613,6 +624,10 @@ public class GroupDialogViewModel {
         return typeTexProperty;
     }
 
+    public BooleanProperty typeSelectionProperty() {
+        return typeSelectionProperty;
+    }
+
     public StringProperty keywordGroupSearchTermProperty() {
         return keywordGroupSearchTermProperty;
     }
@@ -663,5 +678,9 @@ public class GroupDialogViewModel {
 
     public StringProperty texGroupFilePathProperty() {
         return texGroupFilePathProperty;
+    }
+
+    public BooleanProperty entriesAreSelectedProperty() {
+        return entriesAreSelected;
     }
 }